+++ CHANGES.rst
@@ -190,6 +190,15 @@ Bug Fixes
 
 - ``astropy.units``
 
+  - Fixed an  ``ImportError`` that occurred whenever ``astropy.constants`` was
+    imported before ``astropy.units``. [#5030, #5121]
+
+  - Magnitude zero points used to define ``STmag``, ``ABmag``, ``M_bol`` and
+    ``m_bol`` are now collected in ``astropy.units.magnitude_zero_points``.
+    They are not enabled as regular units by default, but can be included
+    using ``astropy.units.magnitude_zero_points.enable()``. This makes it
+    possible to round-trip magnitudes as originally intended.  [#5030]
+
 - ``astropy.utils``
 
 - ``astropy.visualization``